lexigem-app/requirements.txt
Flask
google-generativeai
python-dotenv
Markdown